* Filters no longer additive * Fixed noise gate with sensitivity 0 * Fixed delay filter with delay 0 * Added LED visualisation on clip * Added warning and LED visualisation on long sample processing * Added "volume" and "average" commands * Fixed UI occasionally not showing filter chain

+++ dac.c
@@ -25,8 +25,7 @@ void dac_bias(uint32_t bias)
 }
 
 
-void dac_set(uint32_t val)
+void dac_set(uint16_t val)
 {
 	DAC_UpdateValue(DAC_DEV, val);
 }
-
